Refactor frog still/jumping state (fixes mousewheel)
If the key-up or button-up event arrives immediately after the down event then
rest() was setting the frog back to still before the first jump started.

This had the effect that the scrolling events from a mouse-wheel would create a
frog that was almost immobile.  Scrolling up and down created a pair of frogs
that got stuck on the start line, and since there were two frogs alive the game
didn't restart if all the other frogs died.